fix es pageable search

+++ src/main/java/com/yimingliao/mshivebackend/controller/version1/room/RoomController.java
@@ -41,37 +41,53 @@ public class RoomController {
 
     //-----------------------------------------UPDATE---------------------------------------
     //Update One Room
-    @PatchMapping("/update_one")
-    public R updateOneRoom(@RequestBody Room room) {
+    @PatchMapping("/{user_uuid}/update_one")
+    public R updateOneRoom(@PathVariable("user_uuid") String userUUId, @RequestBody Room room) {
+        if (userService.searchOneUserByUserUUId(userUUId).getStatus() != 200) {
+            return R.error(403, "Insert Forbidden", new Date(), "无权限新增");
+        }
         return roomService.updateOneRoom(room);
     }
 
     //-----------------------------------------DELETE---------------------------------------
     //Delete One User's Room, need roomUUId
-    @DeleteMapping("/delete_one")
-    public R deleteOneRoomByRoomUUId(@RequestParam(name = "roomUUId") String roomUUId) {
+    @DeleteMapping("/{user_uuid}/delete_one")
+    public R deleteOneRoomByRoomUUId(@PathVariable("user_uuid") String userUUId,
+                                     @RequestParam(name = "room_uuid") String roomUUId) {
+        if (userService.searchOneUserByUserUUId(userUUId).getStatus() != 200) {
+            return R.error(403, "Insert Forbidden", new Date(), "无权限新增");
+        }
         return roomService.deleteOneRoomByRoomUUId(roomUUId);
     }
 
     //-----------------------------------------SEARCH---------------------------------------
     //Find One User's Some Rooms, need userUUId, lastSeenRoomId & searchSize
-    @GetMapping("/search_list")
-    public R searchRoomListByUserUUId(@RequestParam(name = "userUUId") String userUUId,
-                                      @RequestParam(name = "lastSeenRoomId") Long lastSeenRoomId,
-                                      @RequestParam(name = "searchSize") Integer searchSize) {
+    @GetMapping("/{user_uuid}/search_list")
+    public R searchRoomListByUserUUId(@PathVariable("user_uuid") String userUUId,
+                                      @RequestParam(name = "last_seen_roomId") Long lastSeenRoomId,
+                                      @RequestParam(name = "search_size") Integer searchSize) {
+        if (userService.searchOneUserByUserUUId(userUUId).getStatus() != 200) {
+            return R.error(403, "Insert Forbidden", new Date(), "无权限新增");
+        }
         return roomService.searchRoomListByUserUUId(userUUId, lastSeenRoomId, searchSize);
     }
 
     //Find One User's All Rooms
-    @GetMapping("/search_all")
-    public R searchRoomAllByUserUUId(@RequestParam(name = "userUUId") String userUUId) {
+    @GetMapping("/{user_uuid}/search_all")
+    public R searchRoomAllByUserUUId(@PathVariable("user_uuid") String userUUId) {
+        if (userService.searchOneUserByUserUUId(userUUId).getStatus() != 200) {
+            return R.error(403, "Insert Forbidden", new Date(), "无权限新增");
+        }
         return roomService.searchRoomAllByUserUUId(userUUId);
     }
 
     //Find One User's One Room, need userUUId & roomId
-    @GetMapping("/search_one")
-    public R searchOneRoomByUserUUId(@RequestParam(name = "userUUId") String userUUId,
-                                     @RequestParam(name = "roomId") Long roomId) {
+    @GetMapping("/{user_uuid}/search_one")
+    public R searchOneRoomByUserUUId(@PathVariable("user_uuid") String userUUId,
+                                     @RequestParam(name = "room_id") Long roomId) {
+        if (userService.searchOneUserByUserUUId(userUUId).getStatus() != 200) {
+            return R.error(403, "Insert Forbidden", new Date(), "无权限新增");
+        }
         return roomService.searchOneRoomByUserUUId(userUUId, roomId);
     }
